Upd 25

+++ MoonBookWeb/DAL/Entities/BookRating.cs
@@ -3,7 +3,7 @@
     public class BookRating
     {
         public Guid Id { get; set; }
-        public int Grade { get; set; }
+        public double Grade { get; set; }
         public Guid IdBook { get; set; }
         public Guid IdUser { get; set; }
     }
